==> Step1_CollectData_Jinmei_FakeNews.ipynb <==
# Scraping news from fake news websites

Fake news websites are taken from
https://github.com/BigMcLargeHuge/opensources/blob/master/sources/sources.csv

## Fake news website list

url_interestingdailynews = 'https://www.interestingdailynews.com'
url_nbccomco = 'http://nbc.com.co'
url_newsbbcnet = 'http://www.newsbbc.net'
url_madworldnews = 'http://madworldnews.com'
url_majorthoughts = 'http://majorthoughts.com'
url_nationonenews = 'http://nationonenews.com'

import newspaper
import pandas as pd
import numpy as np
import os

# collect all or most recent 100 articles from a website
def generate_data(news_url):
    col_names = ["url","source", "title", "author", "text"]
    article_df = pd.DataFrame(columns = col_names)
    
    print(news_url)
    news_articles = newspaper.build(news_url, memoize_articles=False)
    
    size_articles = news_articles.size()
    num_news = 100
    if size_articles < num_news:
        num_news = size_articles
     
    news_brand = news_articles.brand
    count = 0
    for i in range(0,num_news):
        article = news_articles.articles[i];
        try:
            article.download()
            article.parse()
            print(article.url)
            
            entry = pd.DataFrame([[article.url, news_brand, \
                                   article.title, article.authors, article.text]], columns=col_names)
            article_df = article_df.append(entry)
            count += 1
        except:
            pass
        
    print("The total number of " + str(news_brand) + " articles is ", count) 
        
    article_df.set_index('url', inplace=True)
    return article_df    

# check the number of articles in the website before scraping 
# some websites do not have many articles
url = 'http://nationonenews.com'
articles_list = newspaper.build(url,memoize_articles=False)
articles_list.size()

## 1. Obtain news from 'https://www.interestingdailynews.com'

df_interestingdailynews = generate_data(url_interestingdailynews)

output_path = os.path.join('data','indv_file', 'interestingdailynews.csv')
df_interestingdailynews.to_csv(output_path)

# read data
input_path = output_path
data_interestingdailynews = pd.read_csv(input_path)

data_interestingdailynews.head()

## 2. Obtain news from 'http://nbc.com.co'

df_nbccomco = generate_data(url_nbccomco)

path = os.path.join('data','indv_file', 'nbccomco.csv')
df_nbccomco.to_csv(path)  

# test: read data
data_nbccomco = pd.read_csv(path)
data_nbccomco.tail()

## 3. Obtain news from 'http://www.newsbbc.net'

df_newsbbcnet = generate_data(url_newsbbcnet)

path = os.path.join('data', 'indv_file','newsbbcnet.csv')
df_newsbbcnet.to_csv(path)  

# test: read data
data_newsbbcnet = pd.read_csv(path)
data_newsbbcnet.tail()  

## 4. Obtain news from 'http://madworldnews.net'

df_madworldnews = generate_data(url_madworldnews) 

path = os.path.join('data', 'indv_file','madworldnews.csv')
df_madworldnews.to_csv(path)  

# test: read data
data_madworldnews = pd.read_csv(path)
data_madworldnews.tail() 

## 5. Obtain news from 'http://majorthoughts.com'

df_majorthoughts = generate_data(url_majorthoughts)

path = os.path.join('data', 'indv_file','majorthoughts.csv')
df_majorthoughts.to_csv(path)
# test: read data
data_majorthoughts = pd.read_csv(path)
data_majorthoughts.tail()  

## 6. Obtain news from 'http://nationonenews.com'

df_nationonenews = generate_data(url_nationonenews)

path = os.path.join('data','indv_file', 'nationonenews.csv')
df_nationonenews.to_csv(path)
# test: read data
data_nationonenews = pd.read_csv(path)
data_nationonenews.tail() 

## 7. infostormer

url_infostormer = 'http://www.infostormer.com'
df_infostormer = generate_data(url_infostormer)

path = os.path.join('data', 'indv_file','infostormer.csv')
df_infostormer.to_csv(path)
# test: read data
data_infostormer = pd.read_csv(path)
data_infostormer.tail() 

## 8. newswithviews

url_newswithviews = 'http://www.newswithviews.com'
df_newswithviews = generate_data(url_newswithviews)

df_newswithviews.shape

df_newswithviews = \
df_newswithviews.drop(['http://nwpr.org/post/oregon-governor-issues-executive-order-meant-solidify-states-sanctuary-status', \
                       'http://www.nbcnews.com/feature/college-game-plan/mental-health-problems-rising-among-college-students-n777286', \
                       'http://www.washingtontimes.com/news/2017/mar/22/us-heading-toward-lawlessness/',\
                       'http://www.foxnews.com/politics/2017/06/29/house-passes-kate-s-law-as-part-illegal-immigrant-crackdown.html', \
                       'http://www.oregonlive.com/health/index.ssf/2017/07/top_10_oregon_counties_for_gon.html',\
                       'http://www.lifenews.com/2017/02/16/oregon-senate-bill-494-would-allow-starving-mentally-ill-patients-to-death/',\
                       'http://theredelephants.com/breaking-trump-activist-stabbed-9-times/',\
                       'https://en.wikipedia.org/wiki/Civil_forfeiture_in_the_United_States', \
                       'https://www.yahoo.com/news/us-rolls-travel-ban-hit-162422613.html', \
                       'http://www.mailtribune.com/news/20170424/talent-voters-to-decide-on-diesel-tax', \
                       'http://www.mailtribune.com/article/20120724/news/207240315'
                      ])

df_newswithviews.shape

path = os.path.join('data', 'indv_file','newswithviews.csv')
df_newswithviews.to_csv(path)
# test: read data
data_newswithviews = pd.read_csv(path)
data_newswithviews.tail() 

# check the number of articles in the website before scraping 
# some websites do not have many articles
url = 'http://politicot.com'
articles_list = newspaper.build(url,memoize_articles=False)
articles_list.size()

## 9. nephef

url_nephef = 'http://www.nephef.com'
df_nephef = generate_data(url_nephef)

path = os.path.join('data','indv_file', 'nephef.csv')
df_nephef.to_csv(path)
# test: read data
data_nephef = pd.read_csv(path)
data_nephef.tail() 

## 10. local31news

url_local31news = 'http://local31news.com'
df_local31news = generate_data(url_local31news)

path = os.path.join('data','indv_file', 'local31news.csv')
df_local31news.to_csv(path)
# test: read data
data_local31news = pd.read_csv(path)
data_local31news.tail() 

## 11. lastdeplorables

url_lastdeplorables = 'http://lastdeplorables.com'
df_lastdeplorables = generate_data(url_lastdeplorables)

path = os.path.join('data','indv_file', 'lastdeplorables.csv')
df_lastdeplorables.to_csv(path)
# test: read data
data_lastdeplorables = pd.read_csv(path)
data_lastdeplorables.tail() 

## 12. ladylibertysnews

url_ladylibertysnews = 'http://ladylibertysnews.com'
df_ladylibertysnews = generate_data(url_ladylibertysnews)

path = os.path.join('data', 'indv_file','ladylibertysnews.csv')
df_ladylibertysnews.to_csv(path)
# test: read data
data_ladylibertysnews = pd.read_csv(path)
data_ladylibertysnews.tail() 

## 13. krbcnews

url_krbcnews = 'http://krbcnews.com'
df_krbcnews = generate_data(url_krbcnews)

path = os.path.join('data', 'indv_file','krbcnews.csv')
df_krbcnews.to_csv(path)
# test: read data
data_krbcnews = pd.read_csv(path)
data_krbcnews.tail() 

## 14. now8news

url_now8news = 'http://now8news.com' 
df_now8news = generate_data(url_now8news)

path = os.path.join('data','indv_file', 'now8news.csv')
df_now8news.to_csv(path)
# test: read data
data_now8news = pd.read_csv(path)
data_now8news.tail() 

## Merge all data

merge_df = pd.concat([data_interestingdailynews, data_infostormer, 
                      data_krbcnews, 
                      data_local31news, data_lastdeplorables, data_ladylibertysnews, 
                      data_madworldnews, data_majorthoughts, 
                      data_nbccomco, data_newsbbcnet, 
                      data_nationonenews, data_newswithviews, 
                      data_nephef, data_now8news,
                     ], ignore_index=True)
merge_df.tail()

path = os.path.join('data', 'fakenews_jz.csv')
merge_df.to_csv(path)
# test: read data
data_fakenews = pd.read_csv(path)
data_fakenews.tail()
