==> Step1_CollectData_Jinmei_RealNews.ipynb <==
# Scraping news from real news websites

import newspaper
import pandas as pd
import numpy as np
import os

# collect all or most recent 100 articles from a website
def generate_data(news_url, num_news, print_control):
    col_names = ["url","source", "title", "author", "text"]
    article_df = pd.DataFrame(columns = col_names)
    
    print(news_url)
    news_articles = newspaper.build(news_url, memoize_articles=False)
    
    size_articles = news_articles.size()
    if size_articles < num_news:
        num_news = size_articles
     
    news_brand = news_articles.brand
    count = 0
    for i in range(0,num_news):
        article = news_articles.articles[i];
        try:
            article.download()
            article.parse()
            if print_control:
                print(article.url)
            
            entry = pd.DataFrame([[article.url, news_brand, \
                                   article.title, article.authors, article.text]], columns=col_names)
            article_df = article_df.append(entry)
            count += 1
        except:
            pass
        
    print("The total number of " + str(news_brand) + " articles is ", count) 
        
    article_df.set_index('url', inplace=True)
    return article_df    

website_list = ['http://www.nbcnews.com/', 'http://www.msnbc.com/', 'http://www.cbsnews.com/', 
                'http://abcnews.go.com/', 'http://www.newyorker.com/', 
                'http://www.huffingtonpost.com/', 'http://www.bloomberg.com/']

## 1. nbcnews

url_nbcnews = website_list[0]
df_nbcnews = generate_data(url_nbcnews, 200, True)

df_nbcnews.head()

path = os.path.join('data','indv_file', 'nbcnews.csv')
df_nbcnews.to_csv(path)  

# test
# df_nbcnews[df_nbcnews['url'] == '']

articles_list = newspaper.build(url_nbcnews, memoize_articles=False)

for category in articles_list.category_urls(): 
    print(category)

## 2. msnbc news

url_msnbc = website_list[1]
df_msnbc = generate_data(url_msnbc, 200, True)

df_msnbc.shape

path = os.path.join('data', 'indv_file','msnbc.csv')
df_msnbc.to_csv(path)  

## 3. cbsnews

url_cbsnews = website_list[2]
df_cbsnews = generate_data(url_cbsnews, 200, True)

df_cbsnews.shape

path = os.path.join('data','indv_file', 'cbsnews.csv')
df_cbsnews.to_csv(path)  

## 4. abcnews

url_abcnews = website_list[3]
df_abcnews = generate_data(url_abcnews, 200, True)

df_abcnews.shape

path = os.path.join('data', 'indv_file','abcnews.csv')
df_abcnews.to_csv(path)  

## 5. newyorker

url_newyorker = website_list[4]
df_newyorker = generate_data(url_newyorker, 200, True)

df_newyorker.head()

path = os.path.join('data', 'indv_file','newyorker.csv')
df_newyorker.to_csv(path)  

## Merge data

merge_df = pd.concat([df_nbcnews, df_msnbc, df_cbsnews, df_abcnews, df_newyorker])
merge_df.tail()

merge_df.shape

merge_df = merge_df[merge_df.url.str.contains("video") == False]
merge_df.shape

merge_df = merge_df[merge_df.url.str.contains("picture") == False]
merge_df.shape

path = os.path.join('data', 'realnews_jz.csv')
merge_df.to_csv(path)
# test: read data
data_realnews = pd.read_csv(path)
data_realnews.tail()
