==> Step3_MergeData.ipynb <==
# Merge data
* Data format is following Hyesoo's format
* All data is exported to a csv file

import pandas as pd
import os

## Import Jinmei's data

### import fake news

path = os.path.join('data', 'fakenews_jz.csv')
df_fakenews = pd.read_csv(path,usecols=[1,2,3,4,5])
df_fakenews['authenticity'] = 1

### import real news

path = os.path.join('data', 'realnews_jz.csv')
df_realnews = pd.read_csv(path,usecols=[0,1,2,3,4])
df_realnews['authenticity'] = 0

### merge fake and real news

df_jinmei = pd.concat([df_fakenews, df_realnews], ignore_index=True)
df_jinmei.tail()

df_jinmei.shape

### clean data

# drop rows where text contains NaN
df_jinmei = df_jinmei.dropna(axis=0,subset=['text'], how='any') 
df_jinmei.tail()

df_jinmei.shape

# drop rows where the length of characters in text is less than 500
df_jinmei = df_jinmei[df_jinmei.text.map(len) > 500]
df_jinmei.shape

## Import Hyesoo's data

path = os.path.join('data', 'hyesoo_df.csv')
df_hyesoo = pd.read_csv(path,usecols=[1,2,3,4,5,6]) 
swap = {1:0, 0:1}
df_hyesoo.authenticity = df_hyesoo.authenticity.map(swap)
df_hyesoo.tail()

## Import additional fake data

path = os.path.join('data', 'Additional_fake_df.csv')
df_add = pd.read_csv(path,usecols=[1,2,3,4,5,6]) 
df_add.drop(df_add.index[474], inplace = True)
# df_add.tail()
df_add.authenticity.value_counts()

## Merge all data

df_merge = pd.concat([df_jinmei, df_hyesoo, df_add], ignore_index=True)
df_merge.tail()

df_merge.authenticity.value_counts()

## Export data to csv file

path = os.path.join('data','merged_data.csv')
df_merge.to_csv(path)  

A = 1473/1566
A

## Balance data

### sample real news
number of real news ~= number of fake news

data_balanced_realnews = df_merge[df_merge.authenticity==0]
data_balanced_realnews = data_balanced_realnews.sample(frac=0.9406)
data_balanced_realnews.shape

### get all fake news

data_balanced_fakenews = df_merge[df_merge.authenticity==1]
data_balanced_fakenews.shape

### combine real and fake news

data_balanced = pd.concat([data_balanced_realnews, data_balanced_fakenews], ignore_index=True)
data_balanced.authenticity.value_counts()
data_balanced = data_balanced.sample(frac=1).reset_index(drop=True)

### export data

path = os.path.join('data','balanced_data.csv')
data_balanced.to_csv(path)  

data_balanced.head(n=15)
